Move transaction results data structures out of bank.rs
