+++ runtime/src/accounts.rs
@@ -13,13 +13,14 @@ use {
         },
         accounts_update_notifier_interface::AccountsUpdateNotifier,
         ancestors::Ancestors,
-        bank::{Bank, TransactionCheckResult, TransactionExecutionResult},
+        bank::Bank,
         blockhash_queue::BlockhashQueue,
         nonce_info::{NonceFull, NonceInfo},
         rent_collector::RentCollector,
         rent_debits::RentDebits,
         storable_accounts::StorableAccounts,
         transaction_error_metrics::TransactionErrorMetrics,
+        transaction_results::{TransactionCheckResult, TransactionExecutionResult},
     },
     dashmap::DashMap,
     itertools::Itertools,
@@ -1481,8 +1482,8 @@ mod tests {
     use {
         super::*,
         crate::{
-            bank::{DurableNonceFee, TransactionExecutionDetails},
             rent_collector::RentCollector,
+            transaction_results::{DurableNonceFee, TransactionExecutionDetails},
         },
         assert_matches::assert_matches,
         solana_address_lookup_table_program::state::LookupTableMeta,
